dcompile: Avoid importing std.regex
On my (fast) Linux box, this brings the build time for the dcompile
executable down from ~1.35 secs to ~0.6 secs with DMD v2.094.2.
For reggae itself, this means a decrease from 3.6 to 2.9 secs for
the 'metabuild' stage (before invoking make).

While at it, perform the -v output parsing of `import` and `file`
lines in a single pass.