Fix #127
Because of the way reggae was designed, it ignored dub's targetPath
since the way it's intended to be used is by creating a separate build
directory and running it from there. However, users may run it from
the root of the project and in that case it makes sense to respect
dub's targetPath since otherwise the behaviour is different than from
when using dub itself.