Delete extra compiler and linker flags for dub packages
The reason these ever existed was because back in the olden days of
using the information from `dub describe` the compiler flags weren't
added so it was customary to add `-w -g -debug` since otherwise they'd
be missing from debug builds despite dub proper adding them.

Now that reggae uses dub as a library this is no longer necessary.