Avoid threading issues to begin with
The binary backend by default spawns threads for every build, and now
we're using it to both calculate the reggafile dependencies
(in which threads won't even help), and do compile the build generator
from the reggaefile. Since the tests run in threads, it won't be
faster and might create the issues that this branch was created to
fix. In order to still have a safety net, the horrible Windows version
that tries again remains.